+++ src/main/java/com/jgm/springPractice/Main.java
@@ -1,0 +1,17 @@
+package com.jgm.springPractice;
+
+import com.jgm.springPractice.config.Config;
+import com.jgm.springPractice.logic.service.SortService;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+
+import java.util.Arrays;
+
+public class Main {
+    public static void main(String[] args) {
+        ApplicationContext context = new AnnotationConfigApplicationContext(Config.class);
+        SortService sortService = context.getBean(SortService.class);
+
+        System.out.println("[result] " + sortService.doSort(Arrays.asList(args)));
+    }
+}
