+++ src/main/java/com/jgm/springPractice/logic/service/SortService.java
@@ -1,0 +1,22 @@
+package com.jgm.springPractice.logic.service;
+
+import com.jgm.springPractice.logic.Sort;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class SortService {
+
+    private final Sort<String> sort;
+
+    public SortService(@Qualifier("bubbleSort") Sort<String> sort) {
+        this.sort = sort;
+        System.out.println("구현체" +  sort.getClass().getName());
+    }
+
+    public List<String> doSort(List<String> list) {
+        return sort.sort(list);
+    }
+}
