+++ src/test/java/com/jgm/springPractice/MainTest.java
@@ -1,9 +1,6 @@
-package com.jgm.javaoop;
+package com.jgm.springPractice;
 
 import org.junit.jupiter.api.Test;
-
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class MainTest {
